// Soe.Angular/TypeScript/Common/Models/PaymentRowDTO.ts
import { IPaymentRowSaveDTO, IPaymentRowDTO } from "../../Scripts/TypeLite.Net4";
import { AccountingRowDTO } from "./AccountingRowDTO";
import { SoeEntityState, TermGroup_BillingType, TermGroup_InvoiceVatType, SoeOriginType, SoeOriginStatus, SoeInvoiceType, SoePaymentStatus } from "../../Util/CommonEnumerations";

export class PaymentRowDTO implements IPaymentRowDTO {
    paymentRowId: number;
    paymentId: number;
    invoiceId: number;
    paymentImportId: number;
    voucherHeadId: number;
    sysPaymentTypeId: number;
    seqNr: number;
    currencyRate: number;
    currencyDate: Date;
    amount: number;
    amountCurrency: number;
    amountEntCurrency: number;
    amountLedgerCurrency: number;
    amountDiff: number;
    amountDiffCurrency: number;
    amountDiffEntCurrency: number;
    amountDiffLedgerCurrency: number;
    bankFee: number;
    bankFeeCurrency: number;
    bankFeeEntCurrency: number;
    bankFeeLedgerCurrency: number;
    hasPendingAmountDiff: boolean;
    hasPendingBankFee: boolean;
    isRestPayment: boolean;
    isSuggestion: boolean;
    payDate: Date;
    paymentNr: string;
    status: number;
    statusMsg: string;
    statusName: string;
    created: Date;
    createdBy: string;
    modified: Date;
    modifiedBy: string;
    state: SoeEntityState;
    text: string;
    voucherSeriesId: number;
    voucherSeriesTypeId: number;
    paymentMethodId: number;
    description: string;
    billingType: TermGroup_BillingType;
    actorId: number;
    currencyId: number;
    voucherDate: Date;
    invoiceTotalAmount: number;
    invoiceTotalAmountCurrency: number;
    vatAmount: number;
    vatAmountCurrency: number;
    voucherHasMultiplePayments: boolean;
    paidAmount: number;
    paidAmountCurrency: number;
    fullyPaid: boolean;
    paymentAccountRows: AccountingRowDTO[];
    originStatus: number;
    originDescription: string;
    transferStatus: number;
    // Extensions
    isSeqNrTabLabelDisplay: boolean = false;
    // Invoice
    vatType: TermGroup_InvoiceVatType;
    vatCodeId: number;
    totalAmount: number;
    totalAmountCurrency: number;
    totalAmountEntCurrency: number;
    totalAmountLedgerCurrency: number;
    remainingAmount: number;
}

export class PaymentRowSaveDTO implements IPaymentRowSaveDTO {
    originId: number;
    invoiceDate: Date;
    paymentDate: Date;
    voucherDate: Date;
    actorId: number;
    currencyRate: number;
    originType: SoeOriginType;
    originStatus: SoeOriginStatus;
    originDescription: string;
    voucherSeriesId: number;
    accountYearId: number;
    invoiceId: number;
    invoiceType: SoeInvoiceType;
    onlyPayment: boolean;
    billingType: TermGroup_BillingType;
    invoiceNr: string;
    totalAmount: number;
    totalAmountCurrency: number;
    vatAmount: number;
    vatAmountCurrency: number;
    paidAmount: number;
    paidAmountCurrency: number;
    currencyId: number;
    currencyDate: Date;
    fullyPayed: boolean;
    paymentMethodId: number;
    seqNr: number;
    sysPaymentTypeId: number;
    paymentNr: string;
    amount: number;
    amountCurrency: number;
    amountDiff: number;
    amountDiffCurrency: number;
    hasPendingAmountDiff: boolean;
    hasPendingBankFee: boolean;
    state: SoeEntityState;
    text: string;
    voucherHeadId: number;
    paymentStatus: SoePaymentStatus;
    paymentRowId: number;
    importDate: Date;
    importFilename: string;
    isSuperSupportSave: boolean;
    isRestPayment: boolean;
}